Refactored `rustcoalescence/algorithms/cuda/*`
